+++ src/api/diaryapi.js
@@ -15,3 +15,18 @@ export const postDiary = async () => {
     console.log(error);
   }
 };
+
+export const patchDiary = async () => {
+  const obj = {
+    diaryId: 0,
+    title: "string",
+    contents: "string",
+    pics: ["",""],
+  };
+  try {
+    const res = await axios.patch(`api/diary`, obj);
+    console.log(res.data);
+  } catch (error) {
+    console.log(error);
+  }
+};
